+++ anchor/network/src/network.rs
@@ -43,26 +43,6 @@ use crate::{
 
 const MAX_TRANSMIT_SIZE_BYTES: usize = 5_000_000;
 
-/// Count the number of matching subnet bits between two hex-encoded subnet strings
-fn count_matching_subnets(our_subnets: &str, their_subnets: &str) -> usize {
-    // Decode both subnet strings
-    let our_bytes = match hex::decode(our_subnets) {
-        Ok(bytes) => bytes,
-        Err(_) => return 0,
-    };
-    let their_bytes = match hex::decode(their_subnets) {
-        Ok(bytes) => bytes,
-        Err(_) => return 0,
-    };
-
-    // Count matching bits using bitwise AND
-    our_bytes
-        .iter()
-        .zip(their_bytes.iter())
-        .map(|(a, b)| (a & b).count_ones() as usize)
-        .sum()
-}
-
 #[derive(Debug, Error)]
 pub enum NetworkError {
     #[error("Unable to listen on address {address}: {source}")]
@@ -557,37 +537,6 @@ impl<R: MessageReceiver> Network<R> {
         }
     }
 
-    /// Record metrics about subnet overlap after successful handshake.
-    fn record_handshake_subnet_match_metrics(
-        &self,
-        peer_id: PeerId,
-        their_metadata: &NodeMetadata,
-    ) {
-        if let Some(our_metadata) = self.node_metadata() {
-            let matching_count =
-                count_matching_subnets(&our_metadata.subnets, &their_metadata.subnets);
-
-            debug!(
-                %peer_id,
-                our_subnets = %our_metadata.subnets,
-                their_subnets = %their_metadata.subnets,
-                node_version = %their_metadata.node_version,
-                matching_subnets = matching_count,
-                "Handshake completed"
-            );
-
-            // Record subnet match count metric
-            if let Ok(gauge_vec) = crate::metrics::HANDSHAKE_SUBNET_MATCHES.as_ref() {
-                let label = &matching_count.to_string();
-                if let Ok(gauge) = gauge_vec.get_metric_with_label_values(&[label]) {
-                    gauge.inc();
-                }
-            }
-        } else {
-            debug!(%peer_id, "Handshake completed");
-        }
-    }
-
     fn handle_handshake_result(&mut self, event: handshake::Event) {
         match event {
             handshake::Event::Completed {
@@ -604,7 +553,10 @@ impl<R: MessageReceiver> Network<R> {
                         .handle_handshake_completed(peer_id, metadata.node_version.clone());
 
                     // Record subnet matching metrics
-                    self.record_handshake_subnet_match_metrics(peer_id, &metadata);
+                    self.swarm
+                        .behaviour()
+                        .handshake
+                        .record_handshake_subnet_match_metrics(peer_id, &metadata);
                 } else {
                     debug!(%peer_id, ?their_info, "Handshake completed without metadata");
                 }
